requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-pinecone
langchain-huggingface
sentence-transformers
PyMuPDF
pinecone-client
openai
